Add devil-prefix-help-command for alternative prefix help UX
Currently, pressing C-h on a devil sequence will invoke
devil-describe-prefix-bindings, which will call describe-bindings
unconditionally, regardless of the value of prefix-help-command (which
is what would get called outside of a devil sequence).

Integrating prefix-help-command directly is challenging, because
prefix-help-command does not typically take an argument, instead using
this-command-keys or similar to parse the prefix.

This commit adds devil-prefix-help-command, which will be called
instead of devil-describe-prefix-bindings if it is set. The
expectation is for the user to set both prefix-help-command and
devil-prefix-help-command if they want the same (or different!) UX for
prefix binding help.